requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.86.0
Pillow==10.0.1
langdetect==1.0.9
flask==2.3.3
aiosqlite
deep-translator
pandas
streamlit
gradio
aiohttp>=3.9.0
psycopg2-binary
requests
setuptools
httpx
lingua
playwright
pydantic
pytz